requirements.txt
discord.py
boto3
paramiko
python-dotenv
